Activation of farm build emulation functionality
Fixes: https://github.com/containers/podman/issues/26822

Existing farm functionality will only build an image for a particular architecture
if there is a node present in the farm that is native to that architecture. Any
podman servers with qemu emulation capability, which theoretically could also
perform the builds, get ignored.

While leveraging the performance advantage of native hardware is the primary design
intent of the farm, there are still situations where the ability to farm out builds
to emulators serves a useful purpose. Even integration testing of farm functionality
itself would easier if the bulk of it could be done without recourse to secondary
machines.

The existing code has been designed to support emulated builds, at least in principle.
However, the relevant sections of code were never fully completed or activated in
earlier versions of the product.

This commit completes, and makes active, this aspect of the design. Emulated builds
will now be performed, but in _limited_ circumstances; namely, when BOTH of the
following conditions are true :
 * There is no native builder for the architecure available in the farm;
 * The architecture concerned has been explicitly specified with the `--platforms` flag.

When emulated builds are scheduled, they will be performed on the first farm node
found that is capable of the build. This behaviour may be over-ridden in one special
cicumstance; should a farm node be explicitly named after a particular architecture,
then it will be prefered for all emulated builds of that architecture. (i.e farm node
called arm64 will be tasked with performing the arm64 emulated build). In farms
performing multiple emulated build, this allows the workload to be shared across
multiple nodes rather than _all_ being performed on the first capable node found.

The principle code amendments made are:
   * To `pkg/domain/infra/tunnel/farm.go` and `pkg/domain/infra/tunnel/runtime.go`:
       A small number of structs and associated member functions have been amended
       to accomodate the handling of emulated platforms in addition to native
       platforms, namely:
  	- func  fetchInfo()            : now returns a list of platforms the farm is
                                         able to build by emulation, as well as those
                                         it can build natively;

  	- type FarmNode struct{}       : Now includes a array of emulatedPlatforms as
                                         well as a array of nativePlatforms;

  	- FarmNodeInspectReport struct : now includes a list of emulatedPlatforms as
                                         well as a list of nativePlatforms.

   * To pkg/farm/farm.go:
      - `jobs` was originally a mapping `builderName --> platform`, but this was a
         design error. Refactored as a mapping `platform --> buildername`;
      - for emulated builds, the builder allocation algorithm was modified to prefer
        nodes that might be named after a platform architecture.

   * To test/e2e/farm_build_test.go:
       Additional test cases added to exercise the above. For the first time, test
       cases can exercise builds of more than one architecture.

   * To docs/source/markdown/podman-farm-build.1.md.in:
       Documentation updated to describe the circumstances under which emulated
       builds are activated.

Signed-off-by: Kevin Crocombe <kevin.crocombe@pegortech.co.uk>